DAO layer implementation

+++ src/main/java/by/bsuir/wtlab2/entity/Topic.java
@@ -1,6 +1,9 @@
 package by.bsuir.wtlab2.entity;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
 
@@ -9,7 +12,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Topic {
-    private int id;
+    private long id;
     private String name;
     private User author;
     private int questionsCount;
